more error handling

+++ src/main/java/com/librarymanagment/springbootlibrary/controller/StudentController.java
@@ -36,15 +36,17 @@ public class StudentController {
         model.addAttribute("student",students);
         return "new_student";
     }
+    //create new students
     @PostMapping("/saveStudent")
-    public String saveStudent(@Valid @ModelAttribute("student") Students student, BindingResult bindingResult){
+    public String saveStudent(@Valid @ModelAttribute("student") Students student, BindingResult bindingResult, Model model){
 
         if(bindingResult.hasErrors()){
-            return "redirect:/newStudent";
-        }
+            model.addAttribute("inputError", true);
+            return "new_student-error";        }
 
         else if(student.getName().isBlank()||student.getDepartment().isBlank()||student.getEmail().isBlank()){
-            return "redirect:/newStudent";
+            model.addAttribute("blankEntryError", true);
+            return "new_student-error";
         }
         else {
             studentService.save(student);
@@ -52,7 +54,28 @@ public class StudentController {
         }
 
     }
+    //save edits made
+    @PostMapping("/saveStudentEdit")
+    public String saveStudentEdit(@Valid @ModelAttribute("student") Students student, BindingResult bindingResult,Model model)
+    {
+        //if any input errors, show message
+        if(bindingResult.hasErrors()){
+            model.addAttribute("inputError", true);
+            return "edit_student-error";
+        }
+        //if inputs are blank, show message
+        else if(student.getName().isBlank()||student.getDepartment().isBlank()||student.getEmail().isBlank()){
+            model.addAttribute("blankEntryError", true);
+            return "edit_student-error";
+        }
+        //otherwise save edits
+        else {
+            studentService.save(student);
+            return "redirect:/showStudents";
+        }
 
+    }
+    //get borrowed books and show them
     @RequestMapping("/borrowedBooks/{id}")
     public String showBooks(@PathVariable(name = "id") Integer id,Model model, Model model2)
     {
@@ -67,24 +90,35 @@ public class StudentController {
         model2.addAttribute("dateInfoLists",dateInformations);
         return "borrowedBooksList";
     }
+    //edit student with model and view
     @RequestMapping("/editStudent/{id}")
     public ModelAndView editStudent(@PathVariable(name = "id") Integer id) {
         ModelAndView mav = new ModelAndView("edit_student");
         Students student = studentService.findStudent(id);
-        //
-
         mav.addObject("student", student);
-
         return mav;
     }
-    @RequestMapping("/deleteStudent/{id}")
-    public String deleteStudent(@PathVariable(name = "id") Integer id) {
-//        List<Book> book = studentService.findStudent(id).getBookList();
-//        book.remove(studentService.findStudent(id));
-        studentService.deleteStudentByID(id);
 
-        return "redirect:/showStudents";
+    @RequestMapping("/deleteStudent/{id}")
+    public String deleteStudent(@PathVariable(name = "id") Integer id)
+    {
+        //if student is borrowing a book then cannot delete student
+        if(!studentService.ifNotBorrowingBook(id))
+        {
+            return "redirect:/showStudents";
+        }
+        // cannot delete student if debt is owed
+        else if (studentService.findStudent(id).getFines()>0)
+        {
+            return "redirect:/showStudents";
+        }
+        //otherwise delete student
+        else {
+            studentService.deleteStudentByID(id);
+            return "redirect:/showStudents";
+        }
     }
+//    @RequestMapping(value={"/finePayment", "/finePayment-error"})
     @RequestMapping("/finePayment")
     public String payFineRedirect(Model model)
     {
@@ -93,13 +127,16 @@ public class StudentController {
         //directs to finePayment.html
         return "finePayment";
     }
+
+    //did not add custom error html page because student list was not loading.
+    //this function handles fee payments
     @PostMapping(value = "/paymentFineSave")
     public String payFineProcessing(@Valid @RequestParam(name = "student_ID") String student_id,
                                     @RequestParam(name = "paymentInput") Double payment)
     {
+        //if blank reload page
         if(student_id.isBlank())
         {
-
             return "redirect:/finePayment";
         }
         //if ID contains character, reload page
@@ -107,15 +144,17 @@ public class StudentController {
             return "redirect:/finePayment";
         }
 
-        // if student does not exist
+        // if student does not exist reload page
         else if(studentService.ifStudentExists(Integer.parseInt(student_id))==false)
         {
             return "redirect:/finePayment";
         }
+        //otherwise calculate fee payment
         else {
             Students student = studentService.findStudent(Integer.parseInt(student_id));
             double remainder = 0;
             remainder = student.getFines() - payment;
+            //if payment is larger than amount due, then set fine equal to zero
             if (remainder < 0) {
                 remainder = 0;
             }
